requirements-dev.txt
boto3==1.7.14
botocore==1.10.14
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.4
django-pagseguro2==2.0.4
django-storages==1.6.6
docutils==0.14
gunicorn==19.8.1
idna==2.6
jmespath==0.9.3
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.2
python-decouple==3.1
pytz==2018.3
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
static3==0.7.0
urllib3==1.22
xmltodict==0.11.0
